Fix: home 상품보기 선택에 따라 페이지네이션 또는 무한스크롤 구현
