+++ src/views/home/home.html
@@ -50,10 +50,11 @@
         <ul class="prod__list"></ul>
         <!-- 페이지네이션 -->
         <nav class="pagination">
-          <button class="pagination__button-left"><</button>
+          <button class="pagination__button-prev"><</button>
           <div class="pagination__count"></div>
-          <button class="pagination__button-right">></button>
+          <button class="pagination__button-next">></button>
         </nav>
+        <div class="bottom"></div>
       </section>
     </main>
